requirements.txt
pyyaml
mkdocs-material 
mkdocs-macros-plugin